Move JsonConverters into namespace Json.JsonConverters
