+++ Assets/Scripts/Animation/AnimationKeyFrame.cs
@@ -1,6 +1,7 @@
 using PAC.DataStructures;
 using PAC.Extensions;
 using PAC.Json;
+using PAC.Json.JsonConverters;
 
 using System.Runtime.Serialization;
